requirements.txt
openai-agents
openai>=1.0.0
chromadb>=0.4.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.2.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0